requirements.txt
uvicorn==0.13.0
fastapi~=0.65.2
cryptography==3.3.2
python-socketio==4.6.1
python-jose==3.3.0
python-multipart==0.0.5
shapely==1.8.4
eventlet==0.33.1
passlib[bcrypt]==1.7.4
psutil==5.9.2
jinja2==3.0.3
aiofiles==0.8.0
imutils==0.5.4
opencv-python==4.7.0.68
python-vlc==3.0.18121
requests==2.28.2
python-dotenv==0.21.1
Pillow==9.4.0
pyinstaller==5.3.0
auto-py-to-exe==2.22.0
grequests==0.6.0